+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## 0.3.3
 
+- Update glob to regex conversion to prevent command name matching directory components 
 - Handle missing rich dependency, prevent potential ImportErrors when rendering output.
 - Fix error in help for logs subcommand